Move files to the modules folder
