+++ src/modules/Carousel.js
